+++ resources/auth.py
@@ -2,8 +2,8 @@ import logging
 import datetime
 from flask_restful import Resource, reqparse
 from flask_jwt_extended import create_access_token
-from database.models.models import User
-from database.models.base_model import BaseModel
+from database.models import User
+from database.base_model import BaseModel
 
 log = logging.getLogger(__name__)
