+++ resources/user.py
@@ -2,8 +2,8 @@ import logging
 from flask import request
 from flask_restful import Resource, reqparse
 from flask_jwt_extended import jwt_required
-from database.models.models import User, Group
-from database.models.base_model import DatabaseActionException
+from database.models import User, Group
+from database.base_model import DatabaseActionException
 from serialize import serialize
 
 
@@ -96,4 +96,5 @@ class UserResource(Resource):
 
         except DatabaseActionException as ex:
             return {'message': 'Internal Service Error', 'error': ex}, 500
+
